+++ CHANGES.md
@@ -1,3 +1,7 @@
+#2015-04-23
+
+* Merged in the flying brick code
+
 #2015-04-21
 
 * Merge in the investment rate output code